Move apis to api/ and fix config file names
Even if the docs talk about using apis/, the generated new APIs seem to
prefer api/, which is expected since the new kubebuilder is trying to
follow the go idiomatic naming style. Move all the APIs into api/.

When in multi-group project, the resource configs have group name in the
file names too.

Signed-off-by: Sunny <darkowlzz@protonmail.com>