Add default values during conversion
During API upgrade with new fields with default values, before the old
version of the object is converted into new, if a client with new API
queries the object, the fields with default values remain empty. Once
the new version is written those default values are added. To provide
the client with default values in such cases, ensure that the conversion
accounts for the default values as well.

Signed-off-by: Sunny <darkowlzz@protonmail.com>